admin: move CruvedProtectedMixin in utils.py
